\name{parent2012}
\alias{parent2012}
\alias{parent2012dict}
\docType{data}
\title{Data from school questionnaire PISA 2012}
\description{
  The english version of parent questionnaire PISA 2012 is avaliable here \code{http://pisa2012.acer.edu.au/downloads/PISA12_Parent_questionnaire.pdf}.

}
\usage{data(school2012)}
\references{
  http://pisa2009.acer.edu.au/downloads.php
}
\author{
 Dataset created by OECD, transformed by Przemyslaw Biecek
}
\seealso{\code{\link{student2009}}, \code{\link{item2009}}, \code{\link{school2009}}}
\examples{

}
\keyword{datasets}
