\name{school2012}
\alias{school2012}
\alias{school2012dict}
\docType{data}
\title{Data from School Questionnaire PISA 2012}
\description{
  The english version of school questionnaire PISA 2012 is avaliable here \code{http://pisa2012.acer.edu.au/downloads/PISA12_School_questionnaire.pdf}.

}
\usage{data(school2012)}
\references{
  http://pisa2012.acer.edu.au/downloads.php
}
\author{
 Dataset created by OECD, transformed by Przemyslaw Biecek
}
\seealso{\code{\link{student2012}}, \code{\link{item2012}}, \code{\link{parent2012}}}
\examples{
}
\keyword{datasets}
