\name{student2012}
\alias{student2012}
\alias{student2012dict}
\alias{student2012weights}
\docType{data}
\title{Data from Student Questionnaire PISA 2012}
\description{
  The english version of student questionnaire PISA 2012 is avaliable here \code{http://pisa2012.acer.edu.au/downloads/PISA12_Student_questionnaire.pdf}.

  Note that this dataset does not contain replicate weights in order to keep the database smaller than 100MB. You will find replicate weights in the object student2012weights.

}
\usage{data(student2012)}
\references{
  http://pisa2012.acer.edu.au/downloads.php
}
\author{
 Dataset created by OECD, transformed by Przemyslaw Biecek
}
\seealso{\code{\link{school2012}}, \code{\link{item2012}}, \code{\link{parent2012}}}
\examples{
}
\keyword{datasets}
