\name{student2012}
\alias{student2012}
\alias{student2012dict}
\docType{data}
\title{Data from student questionnaire PISA 2012}
\description{
  The english version of student questionnaire PISA 2012 is avaliable here \code{http://pisa2012.acer.edu.au/downloads/PISA12_Student_questionnaire.pdf}.

  Not that this dataset is divided into two subsets: student2012OECD and student2012nonOECD in order to keep each database smaller than 100MB.

}
\usage{data(student2012)}
\references{
  http://pisa2012.acer.edu.au/downloads.php
}
\author{
 Dataset created by OECD, transformed by Przemyslaw Biecek
}
\seealso{\code{\link{school2009}}, \code{\link{item2009}}, \code{\link{parent2009}}}
\examples{
}
\keyword{datasets}
